+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- (libcrux-secrets) [#1446](https://github.com/cryspen/libcrux/pull/1446): Integrate valgrind requests when cfg `valgrind_ct_test` is set
+
+### Removed
+
+- (libcrux-secrets) [#1446](https://github.com/cryspen/libcrux/pull/1446): Remove const constructors of secret types
+
 ## [0.0.4] (2026-05-13)
 
 ### Fixed